+++ backend/telekom-analiz-sistemi/telekom/src/main/java/com/pia/telekom/dto/SubscriptionResponse.java
@@ -3,8 +3,8 @@ package com.pia.telekom.dto;
 import java.time.LocalDate;
 
 public record SubscriptionResponse(
-        Long subscriptionId,
-        Long customerId,
+        Integer subscriptionId,
+        Integer customerId,
         String customerFullName,
         ProductResponse product,
         LocalDate startDate,
