+++ backend/telekom-analiz-sistemi/telekom/src/main/java/com/pia/telekom/repository/RechargeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RechargeRepository extends JpaRepository<Recharge, Long> {
-    Page<Recharge> findByCustomer_CustomerId(Long customerId, Pageable pageable);
+public interface RechargeRepository extends JpaRepository<Recharge, Integer> {
+    Page<Recharge> findByCustomer_CustomerId(Integer customerId, Pageable pageable);
 }
